pyproject.toml
[tool.poetry]
name = "eot2024"
version = "0.1.0"
description = "Cleans the seedlists"
authors = ["Bentley Hensel <bentleyhensel@gmail.com>"]
license = "AGPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
validators = "^0.34.0"
pyyaml = "^6.0.2"
requests = "^2.32.3"
aiohttp = "^3.11.7"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
